refactor: App 책임 분리
App파일 내의 로직을 proxy 파일과 pageRender 파일로 분리하여 수정
